# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Node-Fetch API",
    "version": "1.0.0",
    "description": "订阅管理和用户配置 API 自动生成文档\n\n## 功能特性\n- 🔐 用户配置管理\n- 📊 流量统计\n- 🔄 订阅转换\n- 👥 用户管理（管理员功能）\n- 🗄️ KV 存储服务\n\n## 认证说明\n大部分 API 需要通过 `token` 查询参数进行认证。管理员接口需要 `superToken` 参数。\n\n## 生成时间\n2025-07-15T17:35:03.512Z"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API 服务器"
    },
    {
      "url": "http://localhost:8787/api",
      "description": "开发服务器"
    }
  ],
  "paths": {
    "/create/user": {
      "put": {
        "summary": "创建新用户",
        "description": "管理员接口：创建新用户账户，需要提供用户ID和配置信息",
        "tags": [
          "管理员"
        ],
        "parameters": [
          {
            "name": "superToken",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "超级管理员访问令牌",
            "example": "super_secret_token"
          }
        ],
        "responses": {
          "201": {
            "description": "用户创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未授权访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "用户已存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "SuperTokenAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          }
        }
      }
    },
...
    "/config/users/{userId}": {
      "get": {
        "summary": "获取指定用户配置",
        "description": "根据用户ID获取完整的用户配置信息，包括订阅URL、规则配置等",
        "tags": [
          "用户配置"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "用户唯一标识符",
            "example": "user123"
          },
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "用户访问令牌",
            "example": "user_access_token"
          }
        ],
        "responses": {
          "200": {
            "description": "用户配置信息",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserConfig"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未授权访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "用户不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "TokenAuth": []
          }
        ]
      },
      "post": {
        "summary": "创建或更新用户配置",
        "description": "创建或更新用户配置，支持部分更新",
        "tags": [
          "用户配置"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "用户唯一标识符",
            "example": "user123"
          },
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "用户访问令牌",
            "example": "user_access_token"
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未授权访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "TokenAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserConfig"
              }
            }
          }
        }
      },
      "put": {
        "summary": "更新用户配置",
        "description": "创建或更新用户配置，支持部分更新",
        "tags": [
          "用户配置"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "用户唯一标识符",
            "example": "user123"
          },
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "用户访问令牌",
            "example": "user_access_token"
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未授权访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "TokenAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserConfig"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "删除用户配置",
        "description": "删除指定用户的所有配置信息，操作不可逆",
        "tags": [
          "用户配置"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "用户唯一标识符",
            "example": "user123"
          },
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "用户访问令牌",
            "example": "user_access_token"
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未授权访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "TokenAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserConfig": {
        "type": "object",
        "properties": {
          "subscribe": {
            "type": "string",
            "format": "uri",
            "description": "订阅URL地址"
          },
          "accessToken": {
            "type": "string",
            "description": "用户访问令牌"
          },
          "ruleUrl": {
            "type": "string",
            "format": "uri",
            "description": "规则URL地址",
            "nullable": true
          },
          "fileName": {
            "type": "string",
            "description": "配置文件名称",
            "nullable": true
          },
          "multiPortMode": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AreaCode"
            },
            "description": "多端口模式配置"
          },
          "appendSubList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubConfig"
            },
            "description": "附加订阅列表"
          },
          "excludeRegex": {
            "type": "string",
            "description": "排除正则表达式",
            "nullable": true
          }
        },
        "required": [
          "subscribe",
          "accessToken"
        ],
        "description": "用户配置对象"
      },
      "AreaCode": {
        "type": "string",
        "enum": [
          "HK",
          "TW",
          "SG",
          "JP",
          "KR",
          "US",
          "UK",
          "DE",
          "FR"
        ],
        "description": "地区代码"
      },
      "SubConfig": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "format": "uri"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "url",
          "name"
        ],
        "description": "订阅配置"
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "操作是否成功"
          },
          "message": {
            "type": "string",
            "description": "响应消息"
          },
          "data": {
            "type": "object",
            "description": "响应数据",
            "nullable": true
          }
        },
        "required": [
          "success"
        ],
        "description": "成功响应"
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "错误类型"
          },
          "message": {
            "type": "string",
            "description": "错误消息"
          },
          "code": {
            "type": "number",
            "description": "错误代码"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "错误发生时间"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "description": "错误响应"
      },
...
      "CreateUserRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "新用户的唯一标识符",
            "example": "user123"
          },
          "config": {
            "$ref": "#/components/schemas/UserConfig",
            "description": "用户配置信息"
          },
          "yaml": {
            "type": "string",
            "description": "YAML格式的用户配置（与config二选一）"
          }
        },
        "required": [
          "userId"
        ],
        "description": "创建用户请求"
      },
      "CreateUserResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "创建结果消息",
            "example": "User created successfully"
          },
          "userId": {
            "type": "string",
            "description": "创建的用户ID"
          },
          "config": {
            "$ref": "#/components/schemas/UserConfig",
            "description": "创建的用户配置"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          }
        },
        "required": [
          "message",
          "userId",
          "timestamp"
        ],
        "description": "创建用户成功响应"
      }
    },
    "securitySchemes": {
      "TokenAuth": {
        "type": "apiKey",
        "in": "query",
        "name": "token",
        "description": "用户访问令牌"
      },
      "SuperTokenAuth": {
        "type": "apiKey",
        "in": "query",
        "name": "superToken",
        "description": "超级管理员令牌"
      }
    }
  }
}